fix(plugin): 修复服务端插件 API 和 HTTP 层问题
- BlackBoxPro: onDisable 时调用 BlackBoxApi.shutdown() 清理线程池
- BlackBoxApi: 添加 shutdown() 方法
- BlockActions: useItem 移除无效的 sequence 参数
- EntityActions: interactEntityAt 的 targetX/Y/Z 从 Float 改为 Double
- BlackBoxTestCatalog: click_slot 1.12.2 分支参数名统一为 slot/button/mode
- ExecuteHandler: 添加 1MB 请求体大小限制，handleRunTest 改为异步 whenComplete
- ModRelayClient: 添加 try-finally disconnect、inputStream use、readTimeout 溢出保护

+++ plugin/src/main/kotlin/com/blackboxpro/plugin/api/action/EntityActions.kt
@@ -23,7 +23,7 @@ object EntityActions {
 
     fun interactEntityAt(
         player: Player, entityId: Int,
-        targetX: Float, targetY: Float, targetZ: Float,
+        targetX: Double, targetY: Double, targetZ: Double,
         hand: String = "main_hand", sneaking: Boolean = false
     ): CompletableFuture<ResponseMessage> =
         BlackBoxApi.sendAsync(player, "interact_entity_at", JsonObject().apply {
